[etc] Request DTO(쿠폰생성,발행된쿠폰,유저밸러스업데이트)에 테스트용 생성자 주입

+++ src/main/java/kr/hhplus/be/server/interfaces/api/coupon/dto/request/CouponCreateRequest.java
@@ -17,7 +17,7 @@ public class CouponCreateRequest {
 
     private CouponType type;
 
-    private double valueOfType; //type별 들어가는 값 (ex 정률 할인시 10%할인이면 10, 정액할인시 - 금액 1000원 할인)
+    private int valueOfType; //type별 들어가는 값 (ex 정률 할인시 10%할인이면 10, 정액할인시 - 금액 1000원 할인)
 
     private int remainQuantity; //발급 수량
