+++ src/main/java/kr/hhplus/be/server/interfaces/api/user/dto/request/UserUpdateBalanceRequest.java
@@ -5,17 +5,17 @@ import lombok.Getter;
 
 public class UserUpdateBalanceRequest {
     Long id;
-    int updateBalance;
+    long updateBalance;
 
     public Long getId() {
         return id;
     }
 
-    public int getUpdateBalance() {
+    public long getUpdateBalance() {
         return updateBalance;
     }
 
-    public UserUpdateBalanceRequest(Long id, int updateBalance) {
+    public UserUpdateBalanceRequest(Long id, long updateBalance) {
         this.id = id;
         this.updateBalance = updateBalance;
     }
